Refactor PS1 stuff a bit and add some helpers
I changed the shell color codes to avoid having to deal with escaping
problems. Note, I left a bunch of cruft in the PS1-related code because
it's been annoying me lately, and then MacOS changed the default shell
to ZSH, so I may make that switch across my machines.